pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tembench"
version = "0.1.0"
description = "Language-agnostic benchmarking orchestrator: run commands with parameter sweeps, record metrics, estimate complexity, and generate reports."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "TempoBench Contributors" }
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Testing",
  "Topic :: System :: Benchmark",
]
dependencies = [
  "typer>=0.9",
  "PyYAML>=6.0",
  "pandas>=2.0",
  "altair>=5.0",
  "psutil>=5.9",
  "rich>=13.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.10",
  "pandas-stubs>=2.2",
  "pytest>=7.0",
  "ruff>=0.4",
  "types-psutil>=7.0",
  "types-PyYAML>=6.0",
]

[project.urls]
Homepage = "https://github.com/djeada/TempoBench"
Repository = "https://github.com/djeada/TempoBench"
Issues = "https://github.com/djeada/TempoBench/issues"

[project.scripts]
tembench = "tembench.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 140

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"src/tembench/reporting.py" = ["E501"]

[tool.setuptools.package-data]
tembench = ["py.typed", "reporting/assets/*.css", "reporting/assets/*.js"]

[tool.mypy]
python_version = "3.10"
files = ["src", "tests"]
check_untyped_defs = true
warn_unused_configs = true
pretty = true

[tool.pytest.ini_options]
testpaths = ["tests"]
